fix: Cannot create new Artist Gallery because `_owner` field is undefined.
Wix changed something because suddenly (no warning, no release notes) the `artist` argument WIX passed into `Artists_beforeInsert()` in the backend file `data.js` no longer has an `_owner` field.

Revised the validation logic to expect `_owner` to be undefined. Only errors if it happens to have a value and that is not the same as the signed-in `userid`.

Also
* the failure promise gives more information that will be displayed in the client console
* added paranoia test in case `artist` is null/undefined (and to the `Artists_beforeUpdate()` function as well.